+++ tests/graph/test_connector.py
@@ -25,7 +25,12 @@ from mex.common.models import (
     ExtractedOrganization,
     ExtractedOrganizationalUnit,
 )
-from mex.common.types import Identifier, Text, TextLanguage, Validation
+from mex.common.types import (
+    Identifier,
+    Text,
+    TextLanguage,
+    Validation,
+)
 from tests.conftest import DummyData, MockedGraph, get_graph
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -770,7 +775,7 @@ def test_mocked_graph_fetch_rule_items(mocked_graph: MockedGraph) -> None:
                         "stableTargetId": ["StandaloneRule"],
                     }
                 ],
-                "total": 6,
+                "total": 8,
             },
             id="get-all-rules-when-filtering-for-primary-source-mex-editor",
         ),
@@ -796,7 +801,7 @@ def test_mocked_graph_fetch_rule_items(mocked_graph: MockedGraph) -> None:
                         "stableTargetId": ["StandaloneRule"],
                     }
                 ],
-                "total": 6,
+                "total": 8,
             },
             id="get-all-rules-when-filtering-for-primary-source-mex-editor-and-another-primary-source",
         ),
@@ -812,7 +817,7 @@ def test_mocked_graph_fetch_rule_items(mocked_graph: MockedGraph) -> None:
                         "stableTargetId": ["StandaloneRule"],
                     }
                 ],
-                "total": 6,
+                "total": 8,
             },
             id="no-filters",
         ),
@@ -1140,6 +1145,13 @@ def test_mocked_graph_fetch_merged_items(mocked_graph: MockedGraph) -> None:
                                 "entityType": "SubtractiveOrganizationalUnit",
                                 "stableTargetId": ["StandaloneRule"],
                             },
+                            {
+                                "entityType": "WorkflowOrganizationalUnit",
+                                "forbiddenPublishingTarget": [],
+                                "stableTargetId": [
+                                    "StandaloneRule",
+                                ],
+                            },
                         ],
                         "entityType": "MergedOrganizationalUnit",
                         "identifier": "StandaloneRule",
@@ -1180,6 +1192,11 @@ def test_mocked_graph_fetch_merged_items(mocked_graph: MockedGraph) -> None:
                             {
                                 "email": [],
                                 "entityType": "SubtractiveOrganizationalUnit",
+                                "stableTargetId": ["StandaloneRule"],
+                            },
+                            {
+                                "entityType": "WorkflowOrganizationalUnit",
+                                "forbiddenPublishingTarget": [],
                                 "stableTargetId": ["StandaloneRule"],
                             },
                         ],
@@ -1430,6 +1447,13 @@ def test_mocked_graph_fetch_merged_items(mocked_graph: MockedGraph) -> None:
                                         "language": "en",
                                         "value": "Unit 1.6",
                                     },
+                                ],
+                            },
+                            {
+                                "entityType": "WorkflowOrganizationalUnit",
+                                "forbiddenPublishingTarget": [],
+                                "stableTargetId": [
+                                    "bFQoRhcVH5DHUz",
                                 ],
                             },
                         ],